fix test_metrics
